--- src/utils/gtm.ts
@@ -1,50 +1,0 @@
-// Google Tag Manager utilities for React Router integration
-
-declare global {
-  interface Window {
-    dataLayer: Record<string, unknown>[];
-  }
-}
-
-// Initialize dataLayer if it doesn't exist
-if (typeof window !== 'undefined') {
-  window.dataLayer = window.dataLayer || [];
-}
-
-// Track page views for SPA routing
-export const trackPageView = (url: string, title: string) => {
-  if (typeof window !== 'undefined' && window.dataLayer) {
-    window.dataLayer.push({
-      event: 'page_view',
-      page_title: title,
-      page_location: window.location.origin + url,
-      page_path: url,
-    });
-  }
-};
-
-// Track custom events
-export const trackEvent = (eventName: string, parameters: Record<string, unknown> = {}) => {
-  if (typeof window !== 'undefined' && window.dataLayer) {
-    window.dataLayer.push({
-      event: eventName,
-      ...parameters,
-    });
-  }
-};
-
-// Track form submissions
-export const trackFormSubmission = (formName: string, success: boolean = true) => {
-  trackEvent('form_submit', {
-    form_name: formName,
-    success: success,
-  });
-};
-
-// Track button clicks
-export const trackButtonClick = (buttonName: string, location: string) => {
-  trackEvent('button_click', {
-    button_name: buttonName,
-    click_location: location,
-  });
-};
